+++ dirsync/run.py
@@ -2,6 +2,11 @@
 dirsync's functions
 """
 
+from __future__ import print_function
+
+import sys
+import os
+
 from .syncer import Syncer
 
 
@@ -17,5 +22,18 @@ def sync(sourcedir, targetdir, action, **options):
 
 
 def from_cmdline():
-    from .options import ArgParser
-    sync(**vars(ArgParser().parse_args()))
+    from .options import ArgParser, USER_CFG_FILE, DEFAULT_USER_CFG
+
+    # create config file if it does not exist
+    user_cfg_file = os.path.expanduser(USER_CFG_FILE)
+    if not os.path.isfile(user_cfg_file):
+        print('Creating user config file "%s" ...' % user_cfg_file, end=''),
+        f = open(user_cfg_file, 'w').write(DEFAULT_USER_CFG)
+        f.close()
+        print(' Done')
+
+    try:
+        sync(**vars(ArgParser().parse_args()))
+    except Exception, e:
+        sys.stdout.write(str(e) + '\n')
+        sys.exit(2)